Making the default member map public
